Move 'web' to 'lib/content_translator_web'
An optional step in the Phoenix 1.2 -> 1.3 upgrade instructions: https://gist.github.com/chrismccord/71ab10d433c98b714b75c886eff17357#move-your-root-level-web-directory-to-libapp_name_web-and-migrate-to-new-web-namespace

We've got some stuff in lib/content_translator (workers and API classes) that possibly belong in 'web' as well? For now, I only moved the stuff that was already in 'web'.